Radio, VOR and NDB frequencies are now stored as int and expressed in Hz

+++ src/Airfield.cpp
@@ -12,11 +12,10 @@
 #include "Airfield.h"
 #include <cassert>
 
-Airfield::Airfield(const std::string& longName, const std::string& shortName, const std::string& countryCode, const double lat, const double lon, const float alt, const int style, const int rwyDir, const int rwyLen, const double freq, const std::string& descr)
+Airfield::Airfield(const std::string& longName, const std::string& shortName, const std::string& countryCode, const double lat, const double lon, const float alt, const int style, const int rwyDir, const int rwyLen, const int freq, const std::string& descr)
 	: Waypoint(longName, shortName, countryCode, lat, lon, alt, style, descr)
 	, runwayDir(rwyDir)
 	, runwayLength(rwyLen)
 	, radioFreq(freq) {
 	assert(IsAirfield());
 }
-
