+++ src/Waypoint.h
@@ -55,9 +55,9 @@ public:
 	inline const std::string& GetTypeName() const { return TypeName(type); }
 	inline const std::string& GetDescription() const { return description; }
 	inline bool IsAirfield() const { return IsTypeAirfield(type); }
-	inline void SetOtherFrequency(const double freq) { otherFreq = freq; }
+	inline void SetOtherFrequency(const int freq) { otherFreq = freq; }
 	inline bool HasOtherFrequency() const { return otherFreq > 0; }
-	inline double GetOtherFrequency() const { return otherFreq; }
+	inline int GetOtherFrequency() const { return otherFreq; }
 
 	inline static const std::string& TypeName(const WaypointType& type) { return TYPE_NAMES[type]; }
 
@@ -68,7 +68,7 @@ private:
 	std::string country;
 	float altitude; // [m]
 	WaypointType type;
-	double otherFreq; // frequency for VOR NDB or secondary radio frequency for airports
+	int otherFreq; // [Hz] frequency for VOR NDB or secondary radio frequency for airports
 	std::string description;
 	static const std::string TYPE_NAMES[];
 };
